# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aurora Performance Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gallery:
    get:
      operationId: getGallery
      tags: [gallery]
      summary: List user's generated content
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [photo, video, lipsync, ugc, music_video]
        - name: favorited
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryResponse"
  /gallery/{id}:
    get:
      operationId: getGalleryItem
      tags: [gallery]
      summary: Get a single gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Gallery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
        "404":
          description: Not found
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete a gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /gallery/{id}/favorite:
    patch:
      operationId: toggleFavorite
      tags: [gallery]
      summary: Toggle favorite on a gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteToggle"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
...
  /credits/transactions:
    get:
      operationId: getCreditTransactions
      tags: [credits]
      summary: Get user credit transaction history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionHistory"
  /credits/checkout:
    post:
      operationId: checkoutCredits
      tags: [credits]
      summary: Initiate Paystack payment for credit top-up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Paystack checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
...
components:
  schemas:
...
    GalleryItem:
      type: object
      required: [id, type, status, createdAt, creditsUsed]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [photo, video, lipsync, ugc, music_video]
        status:
          type: string
          enum: [queued, processing, completed, failed]
        prompt:
          type: ["string", "null"]
        outputUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        provider:
          type: ["string", "null"]
        creditsUsed:
          type: integer
        isFavorited:
          type: boolean
        metadataJson:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    GalleryResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GalleryItem"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    FavoriteToggle:
      type: object
      required: [favorited]
      properties:
        favorited:
          type: boolean
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    CheckoutInput:
      type: object
      required: [packageId]
      properties:
        packageId:
          type: string
        email:
          type: ["string", "null"]
    CheckoutResult:
      type: object
      required: [authorizationUrl, reference]
      properties:
        authorizationUrl:
          type: string
        reference:
          type: string
    CreditTransaction:
      type: object
      required: [id, amount, type, description, createdAt, balanceAfter]
      properties:
        id:
          type: string
        amount:
          type: integer
        type:
          type: string
          enum: [topup, deduction, bonus, refund]
        description:
          type: string
        reference:
          type: ["string", "null"]
        balanceAfter:
          type: integer
        createdAt:
          type: string
    TransactionHistory:
      type: object
      required: [transactions, total]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/CreditTransaction"
        total:
          type: integer
...